+++ commands/warn.js
@@ -46,6 +46,7 @@ module.exports = {
             else member = message.guild.members.cache.get(member.id);
             if (!member) return;
   if (!reason) return message.lineReply('Please provide a reason.');
+	if(member.id == message.author.id) return message.lineReply("You can't warn yourself!")
 	message.delete()
 let loading = await message.channel.send("<a:loading:939665977728176168> Give me a sec...")
   member.send(`You have been warned in **${message.guild.name}** for the reason: **${reason}**`);
